Added dialog with all locations
DONE: Added Search Button in Home Page's AppBar which is responsible for displaying Alert Dialog with list of all locations. Clicking an arrow in specific row results in changing camera position.

TODO: Add SearchBar so list of locations will be limited to specific ones after typing some words in it.